requirements.txt
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
pytest
pytest-asyncio
python-dotenv
mcp[cli]
fastapi
uvicorn
httpx
fastmcp
requests
jsonschema
PyJWT
numpy>=1.24
scipy
aiohttp
pynacl
google-generativeai
pandas
matplotlib
aiohttp>=3.8,<4
jsonschema>=4.18.0,<5
pynacl>=1.5,<2
